Add files via upload
My project is coherent customer behavior analysis. In this we have 2 datasets first one customer dataset of 2000 customer which include the information about customer like age education occupation income etc. and second one is purchase data set of 500 customer, purchase history customer with 5 brands of chocolate and promotion or not of every brand.
First, we start with segment dataset we analysis the dataset and perform hierarchical clustering and k means clustering, but we find that cluster are not properly define they are merge into one other. So, we do the PCA dataset, and we go with the 80% of variance of pca and make 3 component and then we perform k means clustering and make 4 segment prospers, ambitious, marginalized, and ordinary and make a scatter plot to see the cluster.
Then we go with purchase dataset which have the 500-customer purchase history we start with the EDA part there is missing value we treat it and know data is ready for analysis then do the Descriptive Analysis by Segmentation and segment our purchase data as per customer data and at last we create a purchase probability model by using logistic regression and find the coefficient which is negative through with we get inferences purchase probability decrease with increase in price. At also get how much increase in price according to segment.
At last, we make a model of purchase probability with or without promotion and get to know that with promotion day the dip in shape is less as compared to no promotion day.